Added stuff for using observer pattern.
Set up GameState to be observable. That means it can register observers,
such as mazeDisplay, and then anytime it believes if ahs changed it can
setChanged and notifyObservers that a change has occurred. Set up
MazeDisplay as an Observer, now implements update() which is called when
something it's registered as observing decides to notify.